Make SharedMemoryHandle a class on windows.
This CL is intended to be a refactor and should not introduce any behavior
changes.

Previously, SharedMemoryhandle was typedefed to HANDLE. Making it a class allows
us to add metainformation about the process in which the HANDLE is valid. This
will be used in the future by Chrome's IPC system to automatically duplicate
HANDLEs into their destination process.

BUG=493414, 535028

Review URL: https://codereview.chromium.org/1320783002

Cr-Original-Commit-Position: refs/heads/master@{#350932}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ea2ab75db847e7506f4e9e6f6b370f65cc1bb39