seccomp-bpf-helpers: Fix undefined reference
Compiling a debug build in the chromeos chroot fails because
we can't link libseccomp_bpf_helpers.so. The failure occurs
because both seccomp_bpf and seccomp_bpf_helpers use
SANDBOX_EXPORT but are in different components and so the
symbols are visible but undefined in some circumstances.

BUG=None
TEST=Build

Review-Url: https://codereview.chromium.org/2291783002
Cr-Original-Commit-Position: refs/heads/master@{#415208}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b7865805147f5c616f13adb38448022d12ac97d2