sandbox tests: Adapt test expectations to new CHECK() behavior
Adapt the test expectations to the changes to official CHECK()
introduced by https://crrev.com/2705053002.
After that CL, CHECK() almost always causes a SIGTRAP with the
exception of armv7 code running on a 32-bit arm kernel, where it
causes a SIGBUS. However, the latter case is not interesting for
the sandbox as it is not enabled on 32-bit arm kernels due to the
lack of tsync.

BUG=694652
TEST=Build with is_official_build=true, run out/X/sandbox_linux_unittests

Review-Url: https://codereview.chromium.org/2709203002
Cr-Original-Commit-Position: refs/heads/master@{#452125}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5d04cccc168b54e6d05f7e9efbb180d87583b30b