Do not check threading status when starting seccomp with SeccompLevel::MULTI_THREADED.
There's no need to assert that the process is in fact multi-threaded,
because the MULTI_THREADED level requires TSYNC support. The check is
inexplicably causing a small number of Android clients to crash. Since
this check does not add value, just remove it.

BUG=664613

Review-Url: https://codereview.chromium.org/2818753002
Cr-Original-Commit-Position: refs/heads/master@{#464490}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 326813f1f1db869240d9a6fe1ce67cd4841a2a4f