macOS V2 Sandbox: Prevent the seatbelt client writev() from blocking.
There are crash reports indicating that the seatbelt client might be
blocking when it tries to writes the profile to the file descriptor. The
best fix for this is to set O_NONBLOCK on the file descriptor.

Bug: 797730
Change-Id: Icb1027d79e7ab9141aafc00d96820ad6f0c0e21e
Reviewed-on: https://chromium-review.googlesource.com/847967
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#527217}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a1bc9d0c559b92d7734e87c954cb21f3097eba14