Add AppContainer support to Service Manager.
This CL adds AppContainer support to the Service Manager sandbox code.
At the moment only the GPU sandbox is supported and in order to enable
the sandbox a new member function must be defined on the sandbox delegate
which provides an identifier to generate a unique AppContainer profile.
This CL does not contain the changes to the GPU code which would enable
the AppContainer sandbox, it's only the infrastructure to enable it in a
further patch set.

Bug: 807249
Change-Id: Iada760c0eab4ff3fae9499bf5310624d964581b7
Reviewed-on: https://chromium-review.googlesource.com/911115
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Penny MacNeil <pennymac@chromium.org>
Commit-Queue: James Forshaw <forshaw@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#535834}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8afd3e92925a52b439f9e49eb93e184b0e7515be