Harmonize the sandbox memory limit across platforms.
...to the extent possible. This involves lowering 64-bit Linux' limit from 8
GiB to 4, to match Windows.

Note that in the process of writing a previous version of this CL, we learned
that macOS apparently does not enforce RLIMIT_DATA, so we continue to not have a
limit on that platform. (The same goes for RLIMIT_AS, which we previously knew
about.)

Bug: 853873,459158
Change-Id: I5a2a76b4c98a0862c16ce121572c4b2b0f910b44
Reviewed-on: https://chromium-review.googlesource.com/1130235
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#573557}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 17b41949b785cb505750e0b65e9f441391283ef4