Remove *Kernel32Loaded APIs from TargetServices.
This API can't be relied on as it needs the NtMapViewOfSection
intercept to be in place, which is not always the case.

There were no callers to this API anyway.

Also, tidy up the hardcoded values to make them more readable.

BUG=None

Change-Id: Ib45b524f465a2730f9164eac987ffa8b35181243
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1696526
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#676523}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 323d285f0740b0e0d722d90bf35b1dc45b1ba0c2