Win Sandbox: Optimize the NtCreateSection intercept.
In the case where an approved DLL is being loaded, there is no
point trying to create the section in the target process, because
it will just fail. In this case, just call straight into the
broker for the section to be created there.

This also fixes issue 993450 as it causes the second call to
NtCreateSection during the LoadLibrary of the chrome DLL to
always be sent via the intercept to the broker.

BUG=995338,993450

Change-Id: Ife7c8c6e1b8c43495e2b69baedbb1096b58a94a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1759896
Reviewed-by: James Forshaw <forshaw@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#688344}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9e30a264e5131bfe5a758f8b05694a543fec19c6