Linux sandbox: add integration tests for syscall broker
The syscall broker has unittests in
//sandbox/linux/syscall_broker/broker_process_unittest.cc, but these
test direct IPC to the broker, and don't use a BPF policy and so aren't
integration tests. This CL adds a suite of integration tests that
mirror the unittests.

This makes it easier to test the user notifier syscall broker, which
is even more in need of integration tests since the kernel itself
sends the syscall arguments to the broker, rather than a custom signal
handler.

The integration tests use a new macro BROKER_TEST, built on top of
the existing BPF_TEST and SANDBOX_TEST infrastructure. It runs each
test with each combination of |fast_check_in_client_| = (true, false),
and method of making a syscall (direct IPC, direct syscall,
libc lib functions).

Bug: 1117351
Change-Id: If1c050f55722487fea51cb479558916983c44985
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2360802
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808515}
GitOrigin-RevId: 3c22878abd562cea507bcb03c853fd13e87a1582