Lock down the Network Sandbox.
This locks the network sandbox to only allow what is needed to
function correctly.

It places Network Service process at UNTRUSTED integrity and
USER_LIMITED token level. This token level allows access to the
DNS service. In order for this to work, a few intercepts
must be landed which are:

 - RW access to user_data directory for the disk cache and other
   network files (e.g. TransportSecurity, Network Persistent State,
   Reporting and NEL databases).
 - RO access to registry policy paths needed for DNS, as these
   are ACLed differently from the rest of HKLM hive.
 - RO access to \DEVICE\\NETBT_TCPIP\* needed for network interface
   enumeration.

The sandbox is gated behind "NetworkServiceSandbox" feature flag
and only enabled on Windows 10 and above.

This CL also fixes up one case where sandbox-type was being used to
control network service launch behavior, removes the exception
for strict handle checks that was added for issue 939590, ensures
that unsandboxed network processes continue to launch in a job
object where available on the platform, and removes an
unneeded global.

BUG=841001,939590,1166831

Change-Id: I3387819f439434d0898f7032b99227ba227cde71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2645644
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850388}
GitOrigin-RevId: 38f381d90b40b1cf42b986d7540a2956cecf7f56