Use sandbox delegate to indicate CET compatibility.
On Windows some processes are not CET compatible but most are. This
adds a CetCompatible method to the sandbox delegate to indicate CET
compatibility. Before this CL the renderer was special cased, but
is now controlled by a delegate. CET remains disabled for the renderer.

This CL also adds to the NaCl delegate to return false as
NaClSyscallCSegHook swaps stacks in a non-compliant manner. This is
new behavior but should (partly) fix 1173919.

This change should only affect Windows 20H1 or later running on Intel
11th gen CPUs.

Tests: browser_tests (including NaCl*) on supporting hardware
Bug: 1173919
Change-Id: Ib4d7384c81c21d605bcad47e45a6021995c04a5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2674433
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Mark Seaborn <mseaborn@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851401}
GitOrigin-RevId: 3c9bf7837aa0963fdc4d0cfece7d385aebac082a