Disable cetcompat when --no-sandbox flag is used.
This CL will disable Hardware-enforced Stack Protection if a process
is launched with the --no-sandbox flag is used. If kNoSandbox is used
then the process delegate indicates if it is compatible with CET.

Previously the --no-sandbox flag would break renderers where CET is
normally disabled.

With this CL the browser process and crashpad-handler will still opt
in to CET. The browser because it is started by Windows, and
crashpad-handler as it is started through its own flow (see #1174399).

Tests: browser_tests & mini_installer on supported hardware
Bug: 1174399,1173919
Change-Id: Ib16a9768c0878897872b159590a1072d4f559a3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2689816
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#854657}
GitOrigin-RevId: 993b077a7272b31a546b10a0402eaf0813d13053