Move Win2kThreadPool creation to BrokerServices Init().
Creation of the thread_pool_ in the broker was deferred until the
first process was spun up in case it was a heavy operation. In reality
the thread pool is just a critical section and some empty containers
so creation will be fast and can happen in the broker's ::Init()
method.

Tangentially related to bug:1167464 but won't affect the crashes.

Tests: sbox_integration_tests
Bug: 1167646
Change-Id: Ia5d015cb091e1bd0f9da7bc508ec401f14ce1173
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2747058
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#861755}
GitOrigin-RevId: eaafd675d72b1ba82622d889d15dbe508af3af50