Allow prctl(PR_PAC_RESET_KEYS) in the sandbox.
https://r.android.com/1652994 proposes to change Bionic to call
prctl(PR_PAC_RESET_KEYS) on each new thread. Since this syscall is
not allowed by the Chromium sandbox, we would otherwise crash when
creating a new thread. To work around the issue in current versions
of Chromium, we avoid issuing the syscall if the target SDK level is
less than S. This CL adds the prctl to the allowlist so that we will
be able to upgrade our target SDK level.

Bug: 919548
Change-Id: If418fcfb827695250336f2889589aaaa62617cc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2787704
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Commit-Queue: Peter Collingbourne <pcc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#867426}
GitOrigin-RevId: 6e749e6521fc796d864c6d26190b245aa7c002a5