Linux sandbox syscall broker: use struct kernel_stat
The struct stat used in libc is different (in size and field ordering)
from the structure assumed by the Linux kernel. So, when emulating
system calls, we need to use the struct definition the kernel expects.

This CL adds linux_stat.h that includes definitions of the different
kernel structs.

Change-Id: I53cad35c2251dff0f6b7ea77528cfa58ef3cab4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2780876
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#871767}
GitOrigin-RevId: 784b0fcd8a3ca6bcd3acb9cfd624ec9cbbac2789