Add histogram for localtime_r() proxying
The Linux localtime_r() can't execute in sandboxed processes so it is
proxied back to the browser. There are GPU hang crash reports that
happen to be in ProxyLocaltimeCallToBrowser().

Add a UMA histogram to get an idea how long this proxy call takes in the
wild. It's possible the GPU watchdog is just killing a slow, instead of
hung, GPU process and it happens to be in this function when the
watchdog fires.

Bug: 1077735
Change-Id: I0242af86244d3eaaf9dbfdebe1487ff38664d10d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2956311
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892191}
NOKEYCHECK=True
GitOrigin-RevId: 723e9076fe1142f40f6aa599d50292e1d4b86971