Unregister waits without holding pool_objects_ lock
Some shutdown hangs report RegisterWait waiting on UnRegisterWait
(e.g. crash/df0741d8d8b45337), usually with a lot of media keys threads
also using the thread pool. This is the same threadpool that the sandbox
is using to manage its waits in RegisterWait/UnRegisterWait. This makes
::UnregisterWaitEx slow as the collection of waits is large and we must
block on it before destroying the sandbox ipc shared memory objects
that use these callbacks.

It is not necessary to hold the |pool_objects_| lock while calling the
Windows API - the same handle cannot be added again as it will still
be open, and the sandbox only calls RegisterWait once for a given
handle. This should allow the registering thread to make progress, and
for shutdown tasks to have a greater chance of completing.

Bug: 1207822
Change-Id: Ia222a3cdc7cdd3cd4e16fb245147b8187ac43521
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2994865
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#897203}
NOKEYCHECK=True
GitOrigin-RevId: d6127bd26824eff38609a7bf1285c476aa1cdffd