Experiment for enabling ExtensionPointDisablePolicy in browser process
This change adds code for an experiment where we will enable the
ProcessExtensionPointDisablePolicy in the browser process, only in
cases where there are no legacy third party IMEs installed.

The plan is that with the experiment enabled, the policy will
initially be off unless the device can be opted in per our heuristic,
which will happen for all devices that don't have legacy third party
IMEs installed. The heuristic is checked after the policy is applied
and therefore changes will only be applied to subsequent launches of
the browser.

Bug: 557798
Change-Id: I27b1f19573a8052513abfee11175a980593b0080
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2990015
Commit-Queue: Stefan Smolen <ssmole@microsoft.com>
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#897213}
NOKEYCHECK=True
GitOrigin-RevId: d5556d4690ed675e2f6759f815a47d7f859ae185