Add access check for low IL on desktop for GPU process
GPU process is failing to load user32.dll when launched with a non-
default Desktop. I suspect that the non-default Desktops are using the
default Security Descriptor from CreateDesktop, which does not allow a
Low-IL process access to the Desktop.

This change detects this case in the Browser process by constructing a
Low-IL token and testing if it has access the Desktop. If the Low-IL
token does not have access, launch the GPU at Medium-IL and lower the GPU
process token after user32.dll is loaded.

Bug: 978133
Change-Id: I312eec567da5126478c27ee2e6077ae03333f160
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3163567
Commit-Queue: Stefan Smolen <ssmole@microsoft.com>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: James Forshaw <forshaw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#923655}
NOKEYCHECK=True
GitOrigin-RevId: 4d514a528153ccf7569e81662d81989066d8eddc