Wrap PrintBackendService in sandboxed and unsandboxed hosts
This removes an action-at-a-distance bool from the print backend
service manager which tracked if the most recently requested service
was in a sandbox or not, and instead uses function-local values. In
particular this improves how the service is launched and allows us
to dispense with service_process_host.h here.

In particular: (Sandboxed|Unsandboxed)PrintBackendHost are launched
as a utility process, and immediately vend the real PrintBackendService.
Bookkeeping is added to the print backend manager to keep these alive.
This continues to use a 1:1:1 mapping of printer:host process:service.

Bug: 1210301
Change-Id: I17cbc18cf49c19eb687aebd30531abb41f2948ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3193682
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Alan Screen <awscreen@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927002}
NOKEYCHECK=True
GitOrigin-RevId: 59a0e967a869c933da2b586e2d19c6f04b045fc0