Linux sandbox: Make BrokerFilePermission more serializable
This converts the boolean fields of BrokerFilePermission to a bitset.
This makes it easy to unserialize and deserialize, and makes it
impossible to miss things in serialization when adding a new flag.

This is preparation for sending the broker policy over IPC.

Bug: 996455
Change-Id: Ifa93e9bd2fc445874bf6fbc69a54c9cd90751156
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3224588
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Cr-Commit-Position: refs/heads/main@{#932793}
NOKEYCHECK=True
GitOrigin-RevId: 8382c3368990b24a07d32e3aef33b8886e71a8a0