Linux sandbox: Refactor to use BrokerSandboxConfig
The BrokerPermissionList and BrokerCommandSet are pretty much always
passed around as a pair. Combine them in a struct, and make it simpler
to land the next CL in which the sandbox client can optionally provide
the BrokerSandboxConfig after starting up the broker by passing it over
IPC.

It's named BrokerSandboxConfig instead of BrokerPolicy to avoid
confusion with the similarly named BPF policies, see
https://crrev.com/c/812285.

This also includes some other small refactors, including renaming
BrokerProcess::Init to BrokerProcess::Fork to be more descriptive, and
Broker-side callbacks take the BrokerSandboxConfig as an argument.

Bug: 996455
Change-Id: I8ad80f701f6392edcff7d2b6a7231a4842774906
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3226200
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Matthew Denton <mpdenton@chromium.org>
Cr-Commit-Position: refs/heads/main@{#937048}
NOKEYCHECK=True
GitOrigin-RevId: 18f8df343ebfd261e3d395aacba77960daa462da