[Windows] Use NtQueryInformationProcess for handle closing.
This CL uses the ProcessHandleTable information class to query the list
of handles for the current process when closing the list of handles.
This information class was added in Windows 8 so for Windows 7 the
handle closer still uses the old brute force approach.

Bug: 1270309
Change-Id: I6153d6a173f41343b0ce43132b98e89bd48724c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3330213
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: James Forshaw <forshaw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#953505}
NOKEYCHECK=True
GitOrigin-RevId: 5ac06479d7072da0a3936e6bdcff74dd73f7cc3e